Added onHideSheet method to Sheet props
Summary:
This component was added back in January, but it hasn't been used, and thus no one has tested it in the real world.

It needs a onHideSheet method to unmount this component when the plugin sheet becomes hidden. This allows the sheet to appear and dissapear on button clicks, instead of only appearing when the container component loads.

For the old behavior see the test plan in this diff: D13597251

Reviewed By: jknoxville

Differential Revision: D16443836

fbshipit-source-id: 68feec23338287dbafc846689ea2a1b35be4b2b0