Upgrade deps within range
Summary:
This is the result of a `yarn upgrade`. While this *shouldn't* break
anything, it easily can just as this broke the babel transform
because a previously transitively depended upon library was no
longer included.

It seems like a good idea to stay as close as possible to the upstream of all
the dependencies we have and with the release yesterday we now have one week to
check if the rest works as expected.

allow-large-files

Reviewed By: danielbuechele

Differential Revision: D16579807

fbshipit-source-id: f26d4536abd80f428e94b71e0fa98df612cbd809