Get all image data for iOS
Summary:
We just implemented the sending of the required metadata of the images plugin for the android, but not for iOS. Due to which we get the following error.

{F173807861}

I tried to implement `getAllImageData` first for iOS in which we send all the information in one message for the images plugin in iOS. But there were lot of images and due to which it exceeded the max payload of the rsocket. That is why I had to implement this bit for iOS in a little different way. We fetch all the image Ids first and then loop over it to get image data.

Due to this the android and iOS implementations are different. I plan to change the android part in the next diffs and make it similar to iOS, as it is more scalable.

Reviewed By: jknoxville

Differential Revision: D16606748

fbshipit-source-id: e98c2bd5db7ec247b45a7cde304d4f51053ea6fe