Update Flipper-Folly.podspec included headers
This removes `folly/**/*.h` in `source_files` which causes all headers to be included. Instead we include only the headers needed to build. In my case this caused a conflict with folly's "Zlib.h" getting included instead of the system "zlib.h" (apparently paths are not case sensitive).
Remove duplicates in `public_header_files`, we can just keep `"folly/**/*.h"` to make all headers public.
Remove `preserve_paths`, this is not be needed anymore.