Remove symlink and copy the actual files
Summary: Removes the symlink files and copies the actual files. The OSS build is broken as its not able to access those files.It used to work before as the ship it config used to sync thos files from fbobjc. But that approach is fragile as it broke our OSS build recently, because of this diff D16829348

Reviewed By: jknoxville

Differential Revision: D16891051

fbshipit-source-id: b5146129adc8628d336fbca5936539d2e1131ba2