remove generics from static methods
Summary:
Generics on static methods don't seem to work the way we are using them.
I looked into how react types getDerivedStateFromProps as this is a very similar concept, and they don't seem to type it at all. For that reason I've given up and just type the persistedState passed to our static methods as `Object`.

Reviewed By: passy

Differential Revision: D17152767

fbshipit-source-id: d9c2159be26fa341c5411dd0b40895d4c8ddf71e