Change default deviceId for MacOS
Summary: It appears the certificate provider is generating a deviceId which is then sent down to the flipper client to store (under CONNECTION_CONFIG_FILE). For MacOS, this ends up being "unknown". Yet when we attempt to connect to the flipper app, it expects empty string. This diff ensures that for MacOS, we're not sending down "unknown" for the client to store into its config file.

Reviewed By: danielbuechele

Differential Revision: D17190130

fbshipit-source-id: e0d9825f1f0330a03bcb3cf8d606d5bd9fcc1d7f