Fix outline icon variant
Summary:
Icon names including the -outline suffix get stripped out so the local path becomes the filled variant by default.
But it's only observable in release builds, because they are the only ones that use local icons.
This should fix that by keeping the variant in the local icon path as well.

Reviewed By: danielbuechele

Differential Revision: D17314646

fbshipit-source-id: b8402d85b1cc690caef7870194e60f22d17dca05