Potential fix for console proxying
Summary: My hypothesis is that on hot reloads / `Ctrl+R`s calls to `this.log` from within the proxy would call the proxy itself. So the only truly relevant change here should be the `.bind(this)` removal. But I'm still not 100% sure that this was the reason for the stack overflow error. Let's keep an eye on this.

Reviewed By: jknoxville

Differential Revision: D17664509

fbshipit-source-id: c9e1c5fe32048b04ef0585f7162bd00ad5ccf18a