Show support form in all the cases
Summary: Earlier we used to show the support form only when selected device was android. But this won't work because, I am planning to add a deeplink support for the form, so that the user can directly open flipper from workplace. Because of this we will have to show the litho support form always. In case  when the selected device is not android, we will show an error screen with possible solutions to open the litho support form.

Reviewed By: passy

Differential Revision: D17787686

fbshipit-source-id: 13e15041238d8746f947563bb3c2ea3cec205a5e