Send flipper trace as a separate message
Summary:
This diff sends flipper trace as a separate message so that the form can update the state which can be used for validation. If the user doesn't click on "Flipper Trace" then it will show a validation error.

WWW side changes are here D17810526. Merge this diff only when www is landed.

Reviewed By: jknoxville

Differential Revision: D17812023

fbshipit-source-id: 35f17f9e58fc08cde3928072dc66758b1c507bcb