Show support form option for deeplink users
Summary: This diff shows the support form option for the deeplink users too even if they are not part of GK. As I discovered a bug where the user opened the Flipper app thru deeplink and then the screen showed temporarily till the device connected and then the default plugin was displayed. In this case there is no way to navigate back to the form.

Reviewed By: jknoxville

Differential Revision: D17978648

fbshipit-source-id: 23ae3660e499a959733753b49cc987f2e381bf6b