Move tests to internal repo
Summary:
Some of the tests need to mock behavior of fb-internal dependencies.
These fb deps get swapped out with a babel transform and it's become hard to maintain now that we're mocking them as well.

For simplicity of mocking, moving them out of the public repo, they'll get run on PRs.

Reviewed By: priteshrnandgaonkar

Differential Revision: D18086247

fbshipit-source-id: 001e258e00da67a112cb754e851253e5480e578a