Fix installed plugin check
Summary:
D17877108 inadvertently introduced a bug because `fs` and `fs-extra` behave differently and apparently wrapping `exists` which is already async in another `promisify` ends up raising the result as an exception (in this case throwing "true" or "false" as an error instead of yielding it back). Annoyingly, TypeScript's type system doesn't catch this. :/

This fixes the invocation and restores the install/remove functionality.

Reviewed By: jknoxville

Differential Revision: D18199958

fbshipit-source-id: f5835d8912bfe1191a12a318103bfc1fb7d776a3