Layout plugin: Add option to Expand/collapse recursively
Summary:
Uses a slightly modified version of the deep expansion. In testing, there were a few issues due to the function in Inspector::onElementExpanded is essentially a toggle. I added an optional parameter to override this toggle behavior for child elements, and also flipped the order of root and child actions during collapsing, as the child state otherwise didn't persist properly (due to the component being unmounted, would be my guess?) This change should be non-breaking to other uses of the method that don't use the `deep: true` parameter
close https://github.com/facebook/flipper/issues/223

Reviewed By: passy

Differential Revision: D18225057

fbshipit-source-id: 53e840f07bf648249b5a4b36d115918dba215ff8