Fix the import and export of the createtable plugin
Summary:
Before this diff, plugins made with createTablePlugin would not be exportable as it uses ImmutableJS's Map which is not serializable. In the custom serializer we just handle the normal map, not the immutableJS one. Thus to tackle this I overrode the serializer and deserializer methods to handle it.

Before this diff the funnellogger import didn't work.
{F221607733}

Reviewed By: mweststrate

Differential Revision: D18352298

fbshipit-source-id: 57d4f3e19f38c12a30e75167646ae43ac8690e08